requirements.txt
chromadb==0.4.15
chainlit
google-generativeai==0.3.2
langchain==0.1.11
langchain-google-genai==0.0.9
numpy==1.26.4
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
pypdf==4.1.0
streamlit==1.32.0
dotenv